requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
python-decouple==3.5
pytz==2021.3
requests==2.27.0
sqlparse==0.4.2
stripe==2.64.0
tzdata==2021.5
urllib3==1.26.7
